lab4/requirements.txt
msoffcrypto-tool
aspose-words
